Fix status type conversion in toWorkflowMetadataView method

+++ core/services/workflows/syncer/v2/grpc_workflow_source.go
@@ -357,7 +357,7 @@ func (g *GRPCWorkflowSource) toWorkflowMetadataView(wf *pb.WorkflowMetadata) (Wo
 		WorkflowID:   workflowID,
 		Owner:        ownerBytes,
 		CreatedAt:    wf.GetCreatedAt(),
-		Status:       uint8(statusVal),
+		Status:       statusVal,
 		WorkflowName: wf.GetWorkflowName(),
 		BinaryURL:    wf.GetBinaryUrl(),
 		ConfigURL:    wf.GetConfigUrl(),
